Refactor: separate computation logic from TUI rendering from CLI. Add PreferredLoads struct.
