+++ internal/keycraft/optimise.go
@@ -1,0 +1,52 @@
+package keycraft
+
+import (
+	"io"
+)
+
+// OptimiseInput encapsulates parameters for BLS optimization.
+type OptimiseInput struct {
+	Layout         *SplitLayout
+	LayoutsDir     string
+	Corpus         *Corpus
+	Prefs          *PreferredLoads
+	Weights        *Weights
+	Pinned         *PinnedKeys
+	NumGenerations int
+	MaxTime        int // minutes
+	Seed           int64
+	LogFile        io.Writer
+}
+
+// OptimiseResult contains optimization results.
+type OptimiseResult struct {
+	OriginalLayout *SplitLayout
+	BestLayout     *SplitLayout
+}
+
+// OptimizeLayout performs BLS optimization.
+// This is the pure computation function that doesn't handle I/O or rendering.
+func OptimizeLayout(input OptimiseInput, consoleWriter io.Writer) (*OptimiseResult, error) {
+	// Run optimization
+	best, err := OptimizeLayoutBLS(
+		input.Layout,
+		input.LayoutsDir,
+		input.Corpus,
+		input.Weights,
+		input.Prefs,
+		input.Pinned,
+		input.NumGenerations,
+		input.MaxTime,
+		input.Seed,
+		consoleWriter,
+		input.LogFile,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return &OptimiseResult{
+		OriginalLayout: input.Layout,
+		BestLayout:     best,
+	}, nil
+}
