+++ internal/keycraft/ranking.go
@@ -9,13 +9,11 @@ import (
 // RankingInput encapsulates all configuration for layout ranking computation.
 // All layouts in LayoutsDir are analyzed for normalization, then filtered to LayoutFiles.
 type RankingInput struct {
-	LayoutsDir     string
-	LayoutFiles    []string      // Specific layouts to rank; if empty, all layouts in LayoutsDir
-	Corpus         *Corpus
-	IdealRowLoad   *[3]float64   // Target distribution for row balance metrics
-	IdealFgrLoad   *[10]float64  // Target distribution for finger balance metrics
-	PinkyPenalties *[12]float64  // Penalty weights for pinky off-home positions
-	Weights        *Weights      // Metric weights for weighted scoring
+	LayoutsDir  string
+	LayoutFiles []string        // Specific layouts to rank; if empty, all layouts in LayoutsDir
+	Corpus      *Corpus
+	Prefs       *PreferredLoads // Load preferences (row, finger, pinky penalties)
+	Weights     *Weights        // Metric weights for weighted scoring
 }
 
 // RankingResult provides ranked layouts with normalization statistics.
@@ -30,7 +28,7 @@ type RankingResult struct {
 // It loads layouts, computes statistics, filters, scores, and returns results.
 func ComputeRankings(input RankingInput) (*RankingResult, error) {
 	// Load and analyze all layouts (needed for normalization even if we filter later)
-	analysers, err := LoadAnalysers(input.LayoutsDir, input.Corpus, input.IdealRowLoad, input.IdealFgrLoad, input.PinkyPenalties)
+	analysers, err := LoadAnalysers(input.LayoutsDir, input.Corpus, input.Prefs)
 	if err != nil {
 		return nil, err
 	}
