+++ internal/tui/corpus_bench_test.go
@@ -1,4 +1,4 @@
-package main
+package tui
 
 import (
 	"testing"
