+++ internal/tui/ranking.go
@@ -1,4 +1,4 @@
-package main
+package tui
 
 import (
 	"encoding/csv"
